add slave error handling
