+++ slave/k8s.go
@@ -1,8 +1,12 @@
 package slave
 
 import (
+  "os"
   "log"
+  "fmt"
   "net"
+  "time"
+  "bufio"
 
   "github.com/keita0805carp/cacis/cacis"
 )
@@ -70,8 +74,108 @@ func installMicrok8s() {
 
 func WaitReadyMicrok8s() {
   log.Printf("[Debug] Waiting for ready\n")
-  cacis.ExecCmd("microk8s status --wait-ready", false)
+  _, err := cacis.ExecCmd("microk8s status --wait-ready --timeout 15", false)
+  if err != nil {
+    log.Printf("[Error] Microk8s is NOT Ready in time\n")
+    log.Printf("[Error] Please try restart or reinstall microk8s.\n")
+    cacis.Error(err)
+  }
   log.Printf("[Debug] Microk8s is Ready\n")
+}
+
+func IsClustered() bool {
+  var lines []string
+  file, err := os.Open("/var/snap/microk8s/current/var/kubernetes/backend/cluster.yaml")
+  cacis.Error(err)
+
+  scanner := bufio.NewScanner(file)
+
+  for scanner.Scan() {
+    line := scanner.Text()
+    lines = append(lines, line)
+  }
+  file.Close()
+
+  if len(lines) > 3 {
+    return true
+  }
+  return false
+}
+
+func clustering(listen net.Listener) {
+  log.Printf("[Debug] Start CLUSTERING\n")
+  if IsClustered() {
+    log.Printf("[Debug] This node join to cluster already.")
+    log.Printf("[Debug] 'cacis slave --leave' to leave cluster")
+    return
+  }
+  // Socket
+  conn2master, err := net.Dial("tcp", masterIP+":"+masterPort)
+  cacis.Error(err)
+  log.Printf("[Debug] Request Clustering\n")
+  cLayer := cacis.RequestClustering()
+  packet := cLayer.Marshal()
+  //fmt.Println(packet)
+  conn2master.Write(packet)
+  log.Printf("Requested\n\n")
+  conn2master.Close()
+
+  conn2slave, err := listen.Accept()
+  cacis.Error(err)
+
+  packet = make([]byte, cacis.CacisLayerSize)
+  packetLength, err := conn2slave.Read(packet)
+  cacis.Error(err)
+  log.Printf("[Debug] Read Packet HEADER. len: %d\n", packetLength)
+  cLayer = cacis.Unmarshal(packet)
+  log.Printf("Debug: Read Packet PAYLOAD\n")
+  cLayer.Payload = loadPayload(conn2slave, cLayer.Length)
+
+  log.Printf("\n[Debug] Clustering...\n")
+  result, err := cacis.ExecCmd(string(cLayer.Payload), true)
+  fmt.Println(string(result))
+  cacis.Error(err)
+
+  log.Printf("[Debug] End CLUSTERING\n")
+  conn2slave.Close()
+}
+
+func Unclustering() {
+  if !IsClustered() {
+    log.Printf("[Error]: Single Node Cluster\n")
+    return 
+  }
+  log.Printf("[Debug] Start UNCLUSTERING\n")
+  log.Printf("[Debug] Leaving...\n")
+  cacis.ExecCmd("microk8s leave", true)
+  // Socket
+  conn, err := net.Dial("tcp", masterIP+":"+masterPort)
+  cacis.Error(err)
+  defer conn.Close()
+
+  hostname, err := os.Hostname()
+  cacis.Error(err)
+
+  log.Printf("[Debug] Request Unclustering\n")
+  cLayer := cacis.RequestUnclustering([]byte(hostname))
+  packet := cLayer.Marshal()
+  //fmt.Println(packet)
+  conn.Write(packet)
+  log.Printf("[Debug] End UNCLUSTERING\n")
+}
+
+func labelNode() {
+  log.Printf("[Debug] Node Labeling...\n")
+  hostname, err := os.Hostname()
+  cacis.Error(err)
+  cmd := "microk8s kubectl label nodes " + hostname
+  for key, value := range cacis.NodeLabels {
+    cmd += " " + key + "=" + value
+  }
+  WaitReadyMicrok8s()
+  time.Sleep(time.Second * 3)
+  cacis.ExecCmd(cmd, false)
+  log.Printf("[Debug] Node Labeled\n")
 }
 
 func RemoveMicrok8s() {
