+++ slave/slave.go
@@ -5,11 +5,11 @@ import (
   "fmt"
   "log"
   "net"
-  "time"
   "encoding/json"
 
   "github.com/keita0805carp/cacis/cacis"
   "github.com/keita0805carp/cacis/master"
+  "github.com/keita0805carp/cacis/connection"
 
   "github.com/containerd/containerd"
 )
@@ -23,19 +23,34 @@ const (
 )
 
 func Main() {
-  listen, err := net.Listen("tcp", ":27001")
-  cacis.Error(err)
+  for {
+    log.Printf("[Debug]: Run Main Slave Process\n")
+    cancel := make(chan struct{})
+    listen, err := net.Listen("tcp", ":27001")
+    cacis.Error(err)
 
-  if !cacis.IsCommandAvailable("microk8s") {
-    cacis.CreateTempDir()
-    recieveMicrok8s(listen)
-    installMicrok8s()
-    setupMicrok8s(listen)
+    ssid, pw := connection.GetWifiInfo()
+    connection.Connect(ssid, pw)
+
+    if !cacis.IsCommandAvailable("microk8s") {
+      cacis.CreateTempDir()
+      recieveMicrok8s(listen)
+      installMicrok8s()
+      setupMicrok8s(listen)
+    }
+
+    WaitReadyMicrok8s()
+    clustering(listen)
+    labelNode()
+
+    go connection.UnstableWifiEvent(cancel)
+
+    <- cancel
+
+    Unclustering()
+    connection.Disconnect()
+    WaitReadyMicrok8s()
   }
-
-  WaitReadyMicrok8s()
-  clustering(listen)
-  labelNode()
 }
 
 func setupMicrok8s(listen net.Listener) {
@@ -43,20 +58,6 @@ func setupMicrok8s(listen net.Listener) {
   sortedExportFileName := cacis.SortKeys(componentsList)
   recieveImg(listen, sortedExportFileName)
   importAllImg(componentsList)
-}
-
-func labelNode() {
-  log.Printf("[Debug] Node Label\n")
-  hostname, err := os.Hostname()
-  cacis.Error(err)
-  cmd := "microk8s kubectl label nodes " + hostname
-  for key, value := range cacis.NodeLabels {
-    cmd += " " + key + "=" + value
-  }
-  WaitReadyMicrok8s()
-  time.Sleep(time.Second * 3)
-  cacis.ExecCmd(cmd, false)
-  log.Printf("[Debug] Node Labeled\n")
 }
 
 func recieveComponentsList(listen net.Listener) map[string]string {
@@ -144,60 +145,6 @@ func importAllImg(m map[string]string) {
     importImg(imageRef, targetDir + importFile)
     log.Printf("[Info]  end\n")
     }
-}
-
-func clustering(listen net.Listener) {
-  log.Printf("[Debug] Start CLUSTERING\n")
-  // Socket
-  conn2master, err := net.Dial("tcp", masterIP+":"+masterPort)
-  cacis.Error(err)
-  log.Printf("[Debug] Request Clustering\n")
-  cLayer := cacis.RequestClustering()
-  packet := cLayer.Marshal()
-  //fmt.Println(packet)
-  conn2master.Write(packet)
-  log.Printf("Requested\n\n")
-  conn2master.Close()
-
-  conn2slave, err := listen.Accept()
-  cacis.Error(err)
-
-
-  packet = make([]byte, cacis.CacisLayerSize)
-  packetLength, err := conn2slave.Read(packet)
-  cacis.Error(err)
-  log.Printf("[Debug] Read Packet HEADER. len: %d\n", packetLength)
-  cLayer = cacis.Unmarshal(packet)
-  log.Printf("Debug: Read Packet PAYLOAD\n")
-  cLayer.Payload = loadPayload(conn2slave, cLayer.Length)
-
-  log.Printf("\n[Debug] Clustering...\n")
-  result, err := cacis.ExecCmd(string(cLayer.Payload), true)
-  fmt.Println(string(result))
-  cacis.Error(err)
-
-  log.Printf("[Debug] End CLUSTERING\n")
-  conn2slave.Close()
-}
-
-func Unclustering() {
-  log.Printf("[Debug] Start UNCLUSTERING\n")
-  log.Printf("[Debug] Leaving...\n")
-  cacis.ExecCmd("microk8s leave", true)
-  // Socket
-  conn, err := net.Dial("tcp", masterIP+":"+masterPort)
-  cacis.Error(err)
-  defer conn.Close()
-
-  hostname, err := os.Hostname()
-  cacis.Error(err)
-
-  log.Printf("[Debug] Request Unclustering\n")
-  cLayer := cacis.RequestUnclustering([]byte(hostname))
-  packet := cLayer.Marshal()
-  //fmt.Println(packet)
-  conn.Write(packet)
-  log.Printf("[Debug] End UNCLUSTERING\n")
 }
 
 func recieveFile(conn net.Conn, fileName string) {
